Use String.prototype.padStart, move pad function
