+++ src/index.test.ts
@@ -1450,32 +1450,6 @@ it('normalize an unsupported type throws', function () {
   );
 });
 
-describe('padWithZeroes', function () {
-  it('pads a string shorter than the target length with zeroes', function () {
-    const input = 'abc';
-    expect(sigUtil.padWithZeroes(input, 5)).toStrictEqual(`00${input}`);
-  });
-
-  it('pads an empty string', function () {
-    const input = '';
-    expect(sigUtil.padWithZeroes(input, 4)).toStrictEqual(`0000`);
-  });
-
-  it('returns a string longer than or equal to the target length without modifying it', function () {
-    const input = 'abc';
-    expect(sigUtil.padWithZeroes(input, 3)).toStrictEqual(input);
-  });
-
-  it('throws an error if passed an invalid hex string', function () {
-    const inputs = ['0xabc', 'xyz', '-'];
-    for (const input of inputs) {
-      expect(() => sigUtil.padWithZeroes(input, 3)).toThrow(
-        new Error(`Expected an unprefixed hex string. Received: ${input}`),
-      );
-    }
-  });
-});
-
 it('personalSign and recover', function () {
   const address = '0x29c76e6ad8f28bb1004902578fb108c507be341b';
   const privKeyHex =
